+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- @imtbl/erc721-permissioned-mintable: Preset contract changes: Added `safeMintBatch` and `burnBatch` to support batch operations.
+
 ### Fixed
 
 - @imtbl/passport: Fixed a bug where the guardian popup was not closing if pre-requisite API calls failed.
 
 ### Changed
 
+- @imtbl/erc721-permissioned-mintable: Preset contract changes: Modified mint method to specify token IDs instead of number of tokens to mint.
 - Modules are now exported under their own namespaces instead of being flattened
 
 ### Removed